refactor(util): move spawn arguments functions to their own package
