+++ src/publisher.ts
@@ -2,7 +2,8 @@ import q = require('q');
 import path = require('path');
 import fs = require('fs');
 import logger from './logger';
-import pactUtil, { DEFAULT_ARG, SpawnArguments } from './pact-util';
+import pactUtil from './pact-util';
+import { DEFAULT_ARG, SpawnArguments } from './spawn-arguments';
 import { deprecate } from 'util';
 import pactStandalone from './pact-standalone';
 const checkTypes = require('check-types');
