+++ src/service.ts
@@ -6,7 +6,8 @@ import events = require('events');
 import http = require('request');
 import q = require('q');
 import logger from './logger';
-import pactUtil, { SpawnArguments } from './pact-util';
+import pactUtil from './pact-util';
+import { SpawnArguments } from './spawn-arguments';
 import { ChildProcess } from 'child_process';
 const mkdirp = require('mkdirp');
 const checkTypes = require('check-types');
